Updating tests to use latest version of jest, which broke a lot of things, but was necessary to get proper ts-jest support.
