+++ packages/server/__mocks__/@sendgrid/mail.ts
@@ -1,0 +1,22 @@
+module SendgridMock {
+  class Email {
+    constructor() {
+      // @ts-ignore
+      this.apiKey = null
+    }
+
+    setApiKey(apiKey: any) {
+      // @ts-ignore
+      this.apiKey = apiKey
+    }
+
+    async send(msg: any) {
+      if (msg.to === "invalid@test.com") {
+        throw "Invalid"
+      }
+      return msg
+    }
+  }
+
+  module.exports = new Email()
+}
