--- packages/server/__mocks__/airtable.js
@@ -1,5 +1,0 @@
-function Airtable() {
-  this.base = jest.fn()
-}
-
-module.exports = Airtable
