+++ packages/server/__mocks__/airtable.ts
@@ -1,0 +1,8 @@
+module AirtableMock {
+  function Airtable() {
+    // @ts-ignore
+    this.base = jest.fn()
+  }
+
+  module.exports = Airtable
+}
