+++ packages/server/__mocks__/mssql.ts
@@ -1,0 +1,24 @@
+module MsSqlMock {
+  const mssql: any = {}
+
+  mssql.query = jest.fn(() => ({
+    recordset: [
+      {
+        a: "string",
+        b: 1,
+      },
+    ],
+  }))
+
+// mssql.connect = jest.fn(() => ({ recordset: [] }))
+
+  mssql.ConnectionPool = jest.fn(() => ({
+    connect: jest.fn(() => ({
+      request: jest.fn(() => ({
+        query: jest.fn(() => ({})),
+      })),
+    })),
+  }))
+
+  module.exports = mssql
+}
