--- packages/server/__mocks__/mysql.js
@@ -1,12 +1,0 @@
-const mysql = {}
-
-const client = {
-  connect: jest.fn(),
-  query: jest.fn((query, bindings, fn) => {
-    fn(null, [])
-  }),
-}
-
-mysql.createConnection = jest.fn(() => client)
-
-module.exports = mysql
