+++ packages/server/__mocks__/mysql.ts
@@ -1,0 +1,14 @@
+module MySQLMock {
+  const mysql: any = {}
+
+  const client = {
+    connect: jest.fn(),
+    query: jest.fn((query, bindings, fn) => {
+      fn(null, [])
+    }),
+  }
+
+  mysql.createConnection = jest.fn(() => client)
+
+  module.exports = mysql
+}
