+++ packages/server/src/integrations/tests/microsoftSqlServer.spec.js
@@ -20,7 +20,7 @@ describe("MS SQL Server Integration", () => {
     const response = await config.integration.create({ 
       sql
     })
-    expect(config.integration.client.query).toHaveBeenCalledWith(sql, undefined)
+    expect(config.integration.client.query).toHaveBeenCalledWith(sql, {})
   })
 
   it("calls the read method with the correct params", async () => {
@@ -28,7 +28,7 @@ describe("MS SQL Server Integration", () => {
     const response = await config.integration.read({ 
       sql
     })
-    expect(config.integration.client.query).toHaveBeenCalledWith(sql, undefined)
+    expect(config.integration.client.query).toHaveBeenCalledWith(sql, {})
   })
 
   describe("no rows returned", () => {
